+++ R/vars.R
@@ -7,6 +7,6 @@ globalVariables(".")
 #'
 #' @export
 #'
-nih.institutes <- c("AA", "AG", "AI", "AR", "AT", "CA", "DA", "DC", "DE", "DK", "EB", "ES", "EY", "GM", "HD", "HG", "HG",
+nih_institutes <- c("AA", "AG", "AI", "AR", "AT", "CA", "DA", "DC", "DE", "DK", "EB", "ES", "EY", "GM", "HD", "HG", "HG",
     "HL", "LM", "MD", "MH", "NR", "NS", "RM", "RR", "TR", "TW")