+++ data-raw/projects.R
@@ -1,6 +1,6 @@
 ## projects and project_pis tables
 
-source('data-raw/common_R')
+source('data-raw/common.R')
 
 path <- here('data-raw/downloads/projects')
 
@@ -42,7 +42,7 @@ projects <- projects_tbl |>
          institute = administering_ic) |>
   filter(!is.na(project_num) & !is.na(total_cost)) |>
   filter(!grepl('-', project_num)) |>
-  filter(institute %in% nih.institutes) |>
+  filter(institute %in% nih_institutes) |>
   mutate(institute = as.factor(institute),
          activity = as.factor(activity),
          application_type = as.factor(application_type),
@@ -62,11 +62,13 @@ project_pis <- projects_tbl |>
   select(core_project_num, pi_ids, administering_ic) |>
   rename(project_num = core_project_num,
          institute = administering_ic) |>
-  filter(institute %in% nih.institutes) |>
+  filter(institute %in% nih_institutes) |>
   filter(!grepl('-', project_num)) |>
   select(project_num, pi_ids) |>
   separate_longer_delim(pi_ids, delim = ';') |>
   filter(pi_ids != '') |>
+  rename(pi_id = pi_ids) |>
+  mutate(pi_id = str_trim(pi_id) |> str_replace_all( ' \\(contact\\)', '')) |>
   na.omit() |> unique() |>
   arrange(project_num)